lab2: added Tabs and useEffect hook.
